refcator AtlasGrid + integrate in SearchPage
