+++ Frontend/src/components/Grids/AtlasesGrid/atlasesGrid.module.css
@@ -1,0 +1,29 @@
+.atlasContainer {
+    margin-block: 1%;
+    padding: 1%;
+    overflow: scroll;
+    overflow-x: hidden;
+    height: 65vh;
+    padding-right: 3%;
+  }
+  
+  .atlasContainer::-webkit-scrollbar {
+    width: 0.8em;
+  }
+  
+  .atlasContainer::-webkit-scrollbar-track {
+    background-color: rgba(128, 128, 128, 0.441);
+    border-radius: 1.5em;
+    margin-block: .5em;
+  }
+  
+  .atlasContainer::-webkit-scrollbar-thumb {
+    background-color: #B7C8E9;
+    border-radius: 1.5em;
+    /*border: .25em solid rgba(128, 128, 128, 0.441); */
+  }
+  
+  .atlasContainer::-webkit-scrollbar-thumb:hover {
+    background-color: #90aee979;
+  
+  }
