+++ Frontend/src/components/SearchPageComponents/SearchContent/index.jsx
@@ -7,10 +7,16 @@ import InstitutionCard from '../SearchResultList/InstitutionCard';
 import UserCard from '../SearchResultList/UserCard';
 import ProjectCard from '../SearchResultList/ProjectCard';
 import ResultStatus from '../ResultStatus';
+import LearnMoreModel from 'views/Explore/LearnMoreModel';
+import AtlasResult from 'views/Explore/AtlasResult';
+import LearnMoreAtlas from 'views/Explore/LearnMoreAtlas';
 import { setSeachCategoryInUrl } from 'shared/utils/common/utils';
+import AtlasesGrid from 'components/Grids/AtlasesGrid';
 
 // wrapper component to display the searched items
-function SearchContent({ searchResult, searchCategory, searchedKeyword }) {
+function SearchContent({
+  searchResult, searchCategory, searchedKeyword,
+}) {
   const { path } = useRouteMatch();
 
   const renderSearchResultsList = (listItemWrapper) => (
@@ -40,6 +46,19 @@ function SearchContent({ searchResult, searchCategory, searchedKeyword }) {
         <Route path={setSeachCategoryInUrl(path, 'projects')}>
           {renderSearchResultsList(ProjectCard)}
         </Route>
+        <Route
+          exact
+          path="/sequencer/search/atlases"
+          render={() => <AtlasesGrid atlases={searchResult} searchedKeyword={searchedKeyword} path="/sequencer/search" />}
+        />
+        <Route
+          exact
+          path="/sequencer/search/models"
+          render={() => <p>Hello models!</p>}
+        />
+        <Route exact path="/sequencer/search/models/:id" render={() => <LearnMoreModel />} />
+        <Route exact path="/sequencer/search/atlases/:id/visualization" render={() => <AtlasResult />} />
+        <Route exact path="/sequencer/search/atlases/:id" render={() => <LearnMoreAtlas />} />
       </Switch>
     </>
   );
