+++ Frontend/src/components/SearchPageComponents/SearchTabs/index.jsx
@@ -5,7 +5,7 @@ import { setSeachCategoryInUrl } from 'shared/utils/common/utils';
 const SearchTabs = ({
   value, onChange, searchParams, path,
 }) => {
-  let categories = ['teams', 'institutions', 'users', 'projects'];
+  let categories = ['atlases', 'models', 'teams', 'institutions', 'users', 'projects'];
   categories = categories.map((category) => ({
     label: category.toUpperCase(),
     path: `${setSeachCategoryInUrl(path, category)}?${searchParams}`,
